\name{EPGLM-package}
\alias{EPGLM-package}
\alias{EPGLM}
\docType{package}
\title{
The package computes a Gaussian approximation of a Bayesian probit/logit using expectation propagation (EP).
}
\description{
The package computes a Gaussian approximation of a Bayesian probit/logit using expectation propagation (EP). For the moment the only available prior is a Gaussian. The algorithm is based on the observation of Chopin and Ridgway [2015].
}
\details{
\tabular{ll}{
Package: \tab EPGLM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-08-23\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Author: James Ridgway\cr
Maintainer: James Ridgway <james.ridgway@ensae.fr>
}
\references{
N. Chopin and J. Ridgway. Leave Pima Indians alone: binary regression as a benchmark for Bayesian computation. arxiv preprint <https://arxiv.org/abs/1506.08640>
}
\keyword{ Bayesian binary regression, Expectation Propagation}
